requirements.txt
# Auto-generated requirements.txt
# Generated by project_launcher.py
# Note: versions are intentionally not pinned. Pin after first successful install if needed.

cryptography
numpy<2
onnxruntime
openai
pillow
pyside6
python-pptx
rank-bm25
